requirements.txt
requests>=2.28.0
pandas>=2.0.0
pyyaml>=6.0
pyarrow>=14.0.0
matplotlib>=3.7
seaborn>=0.12
numpy>=1.24
playwright>=1.40.0
streamlit>=1.28.0
groq>=0.4.0
